Minjeong notice 220315

+++ src/main/java/com/kh/firstclass/admin/place/model/vo/Place.java
@@ -14,8 +14,8 @@ public class Place {
 	private int placeNo;
 	private String placeName;
 	private String placeAddress;
-	private int placeLon;
-	private int placeLat;
+	private double placeLon;
+	private double placeLat;
 	private String placeDes;
 	private String placeTags;
 	private String area;
